task-manager-ld/requirements.txt
flask>=2.0.0
python-dotenv>=1.0.0
launchdarkly-server-sdk>=9.0.0
pyngrok>=7.0.0
